real bug identified
